make type annotations and move back to original location
