+++ mindsdb/integrations/handlers/rag_handler/rag_handler.py
@@ -1,7 +1,7 @@
 from typing import Dict, Optional
 
 import pandas as pd
-from integrations.utilities.rag_evaluate import RAGEvaluator
+from integrations.handlers.rag_handler.rag_evaluate import RAGEvaluator
 
 from mindsdb.integrations.handlers.rag_handler.ingest import Ingestor
 from mindsdb.integrations.handlers.rag_handler.rag import QuestionAnswerer
